Move the method to the Module class
