+++ core-bundle/src/Resources/contao/models/PageModel.php
@@ -595,11 +595,11 @@ class PageModel extends Model
 	 *
 	 * @return Collection|PageModel[]|PageModel|null A collection of models or null if there are no pages
 	 *
-	 * @deprecated Deprecated since Contao 4.9, to be removed in Contao 5.0. Use PageModel::findPublishedWithoutGuestsByPid() instead.
+	 * @deprecated Deprecated since Contao 4.9, to be removed in Contao 5.0. Use Module::getPublishedSubpagesWithoutGuestsByPid() instead.
 	 */
 	public static function findPublishedSubpagesWithoutGuestsByPid($intPid, $blnShowHidden=false, $blnIsSitemap=false)
 	{
-		@trigger_error('Using PageModel::findPublishedSubpagesWithoutGuestsByPid() has been deprecated and will no longer work Contao 5.0. Use PageModel::findPublishedWithoutGuestsByPid() instead.', E_USER_DEPRECATED);
+		@trigger_error('Using PageModel::findPublishedSubpagesWithoutGuestsByPid() has been deprecated and will no longer work Contao 5.0. Use Module::getPublishedSubpagesWithoutGuestsByPid() instead.', E_USER_DEPRECATED);
 
 		$time = Date::floorToMinute();
 		$tokenChecker = System::getContainer()->get('contao.security.token_checker');
@@ -615,46 +615,6 @@ class PageModel extends Model
 		}
 
 		return static::createCollectionFromDbResult($objSubpages, 'tl_page');
-	}
-
-	/**
-	 * Find all published pages by their parent ID and exclude pages only visible for guests
-	 *
-	 * @param integer $intPid        The parent page's ID
-	 * @param boolean $blnShowHidden If true, hidden pages will be included
-	 * @param boolean $blnIsSitemap  If true, the sitemap settings apply
-	 *
-	 * @return array<array{page:PageModel,hasSubpages:bool}>|null
-	 */
-	public static function findPublishedWithoutGuestsByPid($intPid, $blnShowHidden=false, $blnIsSitemap=false): ?array
-	{
-		$time = Date::floorToMinute();
-		$tokenChecker = System::getContainer()->get('contao.security.token_checker');
-		$blnFeUserLoggedIn = $tokenChecker->hasFrontendUser();
-		$blnBeUserLoggedIn = $tokenChecker->hasBackendUser() && $tokenChecker->isPreviewMode();
-
-		$arrPages = Database::getInstance()->prepare("SELECT p1.id, EXISTS(SELECT * FROM tl_page p2 WHERE p2.pid=p1.id AND p2.type!='root' AND p2.type!='error_401' AND p2.type!='error_403' AND p2.type!='error_404'" . (!$blnShowHidden ? ($blnIsSitemap ? " AND (p2.hide='' OR sitemap='map_always')" : " AND p2.hide=''") : "") . ($blnFeUserLoggedIn ? " AND p2.guests=''" : "") . (!$blnBeUserLoggedIn ? " AND p2.published='1' AND (p2.start='' OR p2.start<='$time') AND (p2.stop='' OR p2.stop>'$time')" : "") . ") AS hasSubpages FROM tl_page p1 WHERE p1.pid=? AND p1.type!='root' AND p1.type!='error_401' AND p1.type!='error_403' AND p1.type!='error_404'" . (!$blnShowHidden ? ($blnIsSitemap ? " AND (p1.hide='' OR sitemap='map_always')" : " AND p1.hide=''") : "") . ($blnFeUserLoggedIn ? " AND p1.guests=''" : "") . (!$blnBeUserLoggedIn ? " AND p1.published='1' AND (p1.start='' OR p1.start<='$time') AND (p1.stop='' OR p1.stop>'$time')" : "") . " ORDER BY p1.sorting")
-			->execute($intPid)
-			->fetchAllAssoc();
-
-		if (\count($arrPages) < 1)
-		{
-			return null;
-		}
-
-		// Load models into the registry with a single query
-		static::findMultipleByIds(array_map(static function ($row) { return $row['id']; }, $arrPages));
-
-		return array_map(
-			static function (array $row): array
-			{
-				return array(
-					'page' => static::findByPk($row['id']),
-					'hasSubpages' => (bool) $row['hasSubpages'],
-				);
-			},
-			$arrPages
-		);
 	}
 
 	/**
